move IndexNode to context/types.py
