+++ esgpull/context/__init__.py
@@ -1,12 +1,12 @@
 from esgpull.context.solr import (
     HintsDict,
-    IndexNode,
     ResultSearch,
     _distribute_hits_impl,
 )
 from esgpull.context.solr import (
     SolrContext as Context,
 )
+from esgpull.context.types import IndexNode
 
 __all__ = [
     "Context",
